move code to the analyze_conversation and aggregate_suggestions
