+++ front/lib/reinforced_agent/aggregate_suggestions.ts
@@ -1,6 +1,10 @@
 import { getAgentConfigurations } from "@app/lib/api/assistant/configuration/agent";
+import type { LLMStreamParameters } from "@app/lib/api/llm/types/options";
 import type { Authenticator } from "@app/lib/auth";
-import { runReinforcedAnalysis } from "@app/lib/reinforced_agent/run_reinforced_analysis";
+import {
+  buildReinforcedLLMParams,
+  runReinforcedAnalysis,
+} from "@app/lib/reinforced_agent/run_reinforced_analysis";
 import { AgentSuggestionResource } from "@app/lib/resources/agent_suggestion_resource";
 import logger from "@app/logger/logger";
 import type { AgentInstructionsSuggestionType } from "@app/types/suggestions/agent_suggestion";
@@ -36,6 +40,51 @@ You MUST call the tool. If no suggestions survive aggregation, return an empty s
 ${suggestionsSection}`;
 
   return { systemPrompt, userMessage };
+}
+
+/**
+ * Build the batch map for aggregation.
+ * Returns null if there are no pending synthetic suggestions or the agent is not found.
+ */
+export async function buildAggregationBatchMap(
+  auth: Authenticator,
+  agentConfigurationId: string
+): Promise<Map<string, LLMStreamParameters> | null> {
+  const syntheticSuggestions =
+    await AgentSuggestionResource.listByAgentConfigurationId(
+      auth,
+      agentConfigurationId,
+      { sources: ["synthetic"], states: ["pending"] }
+    );
+
+  if (syntheticSuggestions.length === 0) {
+    return null;
+  }
+
+  const [agentConfig] = await getAgentConfigurations(auth, {
+    agentIds: [agentConfigurationId],
+    variant: "light",
+  });
+  if (!agentConfig) {
+    logger.warn(
+      { agentConfigurationId },
+      "ReinforcedAgent: agent not found for aggregation batch"
+    );
+    return null;
+  }
+
+  const instructionsSuggestions = syntheticSuggestions
+    .map((s) => s.toJSON())
+    .filter(
+      (json): json is AgentInstructionsSuggestionType =>
+        json.kind === "instructions"
+    );
+
+  const prompt = buildAggregationPrompt(
+    agentConfig.name,
+    instructionsSuggestions
+  );
+  return new Map([["aggregation", buildReinforcedLLMParams(prompt)]]);
 }
 
 /**
